Support forwarding events to another Slipstream
This works by using two plugins;
 1) TransmitterPlugin will serialize most* incoming events and
    send them via TCP/IP to a receiver, as configured in settings
 2) ReceiverPlugin deserializes the incoming data to events
    and sends them on its own eventbus

*most events are sent: "Internal" events such as Plugin management
(register/unregister/enable/disable) are ignoed as we wouldnt be able to
use different plugins on the receiver side

Filemonitoring events are also ignored. The files are not shared
between, so these events doesnt make much sense on the receiver.

Utility events (writing to console, saying text, playing audio) are also
ignored.

The protocol between Transmitter and Receiver is json based. Each event
is converted to json and each json string gets a linefeed (\n) as
delimiter. json strings are assumed not to contain linefeeds!