Lazily publishing events from lua scripts at start
LuaPlugin will capture all events the script tries to send when they are
initially evaluated. These events are converted to json and sent via an
InternalBootupEvents event. Engine will listen for these, unwrap the
events and add it to an internal pending event list. If no
InternalBootupEvents are received within 0.5s, it will remove duplicates
and publish the events.

This means, that if multiple Lua scripts sends the same event (such as
asking IRacingPlugin to return some specific data), it will be reduced
to a single event, to avoid having the same information sent multiple
times.