Use Autofac to create plugins
LuaGlueFactories are not removed and a part of a Plugin. Upon starting
loading a lua script, it will ask all existing plugins for their glues
(if any) and apply them.

A side-effect of this meant, that Components without Plugins but with
LuaGlue wouldn't work. To fix this, two new plugins were defined:

- InternalPlugin
- UIPlugin

Could finally remove all the component registration code, so now
component is merely a directory containing plugins, factories, events
and handlers.