Reimplement application update using AutoUpdaterDotNET
It seems that Squirrel.Windows is dying. There have been an
vulnerbility[0] for in one of the dependencies, discovered on september
12th, 2019, that can be fixed by merely releasing a new version with
updated dependency. It's a trivial task, yet it wasn't done.

Also it seems that there was an attempt[1], but that didn't really find
new maintainers. electron consider the project as not maintained[2]

0: https://github.com/advisories/GHSA-fxh6-w476-hgr4
1: https://github.com/Squirrel/Squirrel.Windows/issues/1470
2: https://github.com/electron/electron/issues/17722