Bugfix: Events shared timestamp
The timestamp `uptime` is updated when you publish an event. As we only
create a new envelope everytime we changed/add recipients, multiple
events would share the same envelope. This isn't a problem, unless you
consider the envelope also contains the timestamp. So suddenly all
events sharing the same envelope would have the timestamp of the last
event using that envelope.

When viewing the events they would seem out of order.