JustGiving: Dont restart fetching donation if result changes
For busy fundraising, we will struggle to get all donations. So instead
we just read it fully and then pick any new donations up on next check.
Now that we check donations based on donation ids, this should be fine.
Assuming that all new donations are on the first pages