Bugfix: Removing the endpoint menuitem
When a new endpoint is started it will announce
itself as http://*/... but in the menu will be shown
as http://127.0.01/... so you can use it directly.
However when endpoint remove announcement
comes in, we would look for menu with the
original value and not finding any. So we would
never remove endpoints from the menu. Fixed.